utility_service/requirements.txt
# requirements.txt
Flask
Flasgger
requests
mysql-connector-python
python-dotenv 
bcrypt # Untuk hashing password
PyJWT # Untuk membuat token autentikasi JSON